requirements.txt
aiohttp==3.8.1
aiohttp-sse==2.1.0
schematics==2.1.1
asyncpg==0.25.0
PyJWT==2.3.0
cryptography==36.0.1
pgrate==0.1.1
grpcio-tools==1.47.0
